+++ CHANGELOG.md
@@ -49,6 +49,7 @@
   - `XargsCat()` now supports tracing
   - `XargsDirname()` now supports tracing
   - `XargsRmFile()` now supports tracing
+  - `XargsTestFilepathExists()` now supports tracing
 * Pipelines now set a context flag in their Pipe
   - this is used to tell sinks where to read from!